+++ Beam/extensions/beam-module/tests/Feature/CompressAggregationsTest.php
@@ -7,8 +7,6 @@ use Remp\BeamModule\Model\ArticlePageviews;
 use Remp\BeamModule\Model\ArticleTimespent;
 use Remp\BeamModule\Console\Commands\CompressAggregations;
 use Remp\BeamModule\Model\Property;
-use Remp\BeamModule\Model\SessionDevice;
-use Remp\BeamModule\Model\SessionReferer;
 use Carbon\Carbon;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Remp\BeamModule\Tests\TestCase;
@@ -35,91 +33,6 @@ class CompressAggregationsTest extends TestCase
     {
         $this->runAndTestArticleAggregations(ArticleTimespent::class);
     }
-
-    public function testAggregateAllDaySessionReferers()
-    {
-        $start = Carbon::today()->subDays($this->thresholdPeriod + 1);
-        for ($i = 0; $i < 24; $i++) {
-            SessionReferer::factory()->create([
-                'time_from' => (clone $start)->addHours($i),
-                'time_to' => (clone $start)->addHours($i + 1),
-                'subscriber' => false,
-                'count' => 1,
-                'medium' => 'a',
-                'source' => 'a',
-            ]);
-        }
-
-        $this->assertEquals(24, SessionReferer::all()->count());
-
-        $this->artisan(CompressAggregations::COMMAND);
-
-        $this->assertEquals(1, SessionReferer::all()->count());
-    }
-
-    public function testSessionDevicesAggregation()
-    {
-        $same = [
-            'type' => 'a',
-            'model' => 'a',
-            'brand' => 'a',
-            'os_name' => 'a',
-            'os_version' => 'a',
-            'client_type' => 'a',
-            'client_name' => 'a',
-            'client_version' => 'a',
-        ];
-
-        SessionDevice::factory()->create(array_merge($same, [
-                'time_from' => Carbon::today()->hour(9)->subDays($this->thresholdPeriod),
-                'time_to' => Carbon::today()->hour(10)->subDays($this->thresholdPeriod),
-                'subscriber' => false,
-                'count' => 10,
-            ])
-        );
-
-        SessionDevice::factory()->create(array_merge($same, [
-                'time_from' => Carbon::today()->hour(9)->subDays($this->thresholdPeriod),
-                'time_to' => Carbon::today()->hour(10)->subDays($this->thresholdPeriod),
-                'subscriber' => false,
-                'count' => 5,
-            ])
-        );
-
-        SessionDevice::factory()->create(array_merge($same, [
-                'time_from' => Carbon::today()->hour(9)->subDays($this->thresholdPeriod),
-                'time_to' => Carbon::today()->hour(10)->subDays($this->thresholdPeriod),
-                'subscriber' => true,
-                'count' => 2,
-            ])
-        );
-
-        //// Do not aggregate these
-        SessionDevice::factory()->create(array_merge($same, [
-                'time_from' => Carbon::today()->hour(9)->subDays($this->thresholdPeriod - 1),
-                'time_to' => Carbon::today()->hour(10)->subDays($this->thresholdPeriod - 1),
-                'subscriber' => false,
-                'count' => 6,
-            ])
-        );
-
-        SessionDevice::factory()->create(array_merge($same, [
-                'time_from' => Carbon::today()->hour(9)->subDays($this->thresholdPeriod - 1),
-                'time_to' => Carbon::today()->hour(10)->subDays($this->thresholdPeriod - 1),
-                'subscriber' => false,
-                'count' => 3,
-            ])
-        );
-
-        $this->artisan(CompressAggregations::COMMAND);
-
-        $this->assertEquals(4, SessionDevice::all()->count());
-
-        $d = Carbon::today()->subDays($this->thresholdPeriod);
-
-        $this->assertEquals(15, SessionDevice::whereDate('time_from', $d)->where('subscriber', false)->first()->count);
-    }
-
 
     public function runAndTestArticleAggregations($className)
     {